Refactored event manager integration spec into individual files
