+++ scala-debugger-api/src/it/scala/org/senkbeil/debugger/jdi/events/filters/MinTriggerFilterIntegrationSpec.scala
@@ -1,15 +1,13 @@
-package org.senkbeil.debugger.events
+package org.senkbeil.debugger.jdi.events.filters
 
 import com.sun.jdi.event.BreakpointEvent
 import org.scalatest.concurrent.Eventually
-import org.scalatest.time.{Seconds, Milliseconds, Span}
+import org.scalatest.time.{Milliseconds, Seconds, Span}
 import org.scalatest.{FunSpec, Matchers, ParallelTestExecution}
-import org.senkbeil.debugger.jdi.events.filters.{MaxTriggerFilter, MinTriggerFilter}
 import test.{TestUtilities, VirtualMachineFixtures}
+import org.senkbeil.debugger.events.EventType._
 
-import EventType._
-
-class EventManagerIntegrationSpec extends FunSpec with Matchers
+class MinTriggerFilterIntegrationSpec extends FunSpec with Matchers
   with ParallelTestExecution with VirtualMachineFixtures
   with TestUtilities with Eventually
 {
@@ -18,7 +16,7 @@ class EventManagerIntegrationSpec extends FunSpec with Matchers
     interval = scaled(Span(5, Milliseconds))
   )
 
-  describe("EventManager") {
+  describe("MinTriggerFilter") {
     it("should ignore the first N events for a handler using MinTriggerFilter(N)") {
       val testClass = "org.senkbeil.test.filters.MinTriggerFilter"
       val testFile = scalaClassStringToFileString(testClass)
@@ -31,45 +29,6 @@ class EventManagerIntegrationSpec extends FunSpec with Matchers
 
       // Expected breakpoints to invoke handler
       val expected = Seq(12, 13)
-
-      // Will contain the hit breakpoints
-      @volatile var actual = collection.mutable.Seq[Int]()
-
-      withVirtualMachine(testClass, suspend = false) { (v, s) =>
-        // Queue up our breakpoints
-        breakpointLines.foreach(
-          s.breakpointManager.setLineBreakpoint(testFile, _: Int)
-        )
-
-        // Queue up a generic breakpoint event handler that filters events
-        s.eventManager.addResumingEventHandler(BreakpointEventType, e => {
-          val breakpointEvent = e.asInstanceOf[BreakpointEvent]
-          val location = breakpointEvent.location()
-          val fileName = location.sourcePath()
-          val lineNumber = location.lineNumber()
-
-          logger.debug(s"Reached breakpoint: $fileName:$lineNumber")
-          actual :+= lineNumber
-        }, filter)
-
-        logTimeTaken(eventually {
-          actual should contain theSameElementsInOrderAs (expected)
-        })
-      }
-    }
-
-    it("should ignore all events for a handler after the first N using MaxTriggerFilter(N)") {
-      val testClass = "org.senkbeil.test.filters.MaxTriggerFilter"
-      val testFile = scalaClassStringToFileString(testClass)
-
-      // The filter to apply (should ignore breakpoints after first three)
-      val filter = MaxTriggerFilter(count = 3)
-
-      // Mark lines we want to potentially breakpoint
-      val breakpointLines = Seq(9, 10, 11, 12, 13)
-
-      // Expected breakpoints to invoke handler
-      val expected = Seq(9, 10, 11)
 
       // Will contain the hit breakpoints
       @volatile var actual = collection.mutable.Seq[Int]()
